add option to add tags to books. add option to remove tags from books. add auth/me. add option to remove own review.

+++ BookReviewer/Models/DTOs/Books/BookListDto.cs
@@ -6,7 +6,7 @@
         public string ISBN { get; set; } = default!;
         public string Title { get; set; } = default!;
 
-        public double? AverageRating { get; set; }
+        public decimal? AverageRating { get; set; }
         public List<string> Authors { get; set; } = new();
     }
 }
